+++ Source/Bake/Cooking/Composers/MkDocsComposer.cs
@@ -21,7 +21,9 @@
 // SOFTWARE.
 
 using Bake.Core;
+using Bake.Services;
 using Bake.ValueObjects.Artifacts;
+using Bake.ValueObjects.Destinations;
 using Bake.ValueObjects.Recipes;
 using Bake.ValueObjects.Recipes.MkDocs;
 using Bake.ValueObjects.Recipes.Pip;
@@ -34,6 +36,7 @@ namespace Bake.Cooking.Composers
     public class MkDocsComposer : Composer
     {
         private readonly IFileSystem _fileSystem;
+        private readonly IDockerLabels _dockerLabels;
 
         public override IReadOnlyCollection<ArtifactType> Produces { get; } = new[]
             {
@@ -41,9 +44,11 @@ namespace Bake.Cooking.Composers
             };
 
         public MkDocsComposer(
-            IFileSystem fileSystem)
+            IFileSystem fileSystem,
+            IDockerLabels dockerLabels)
         {
             _fileSystem = fileSystem;
+            _dockerLabels = dockerLabels;
         }
 
         public override async Task<IReadOnlyCollection<Recipe>> ComposeAsync(
@@ -55,13 +60,21 @@ namespace Bake.Cooking.Composers
                 "mkdocs.yml",
                 cancellationToken);
 
+            var labels = await _dockerLabels.FromIngredientsAsync(
+                context.Ingredients,
+                cancellationToken);
+            var buildContainer = context.Ingredients.Destinations
+                .OfType<ContainerDestination>()
+                .Any(d => d.ArtifactType == ArtifactType.DocumentationSite);
+
             return mkDocsFilePaths
-                .SelectMany(CreateRecipes)
+                .SelectMany(p => CreateRecipes(p, labels, buildContainer))
                 .ToArray();
         }
 
-        private static IEnumerable<Recipe> CreateRecipes(
-            string mkDocsFilePath)
+        private static IEnumerable<Recipe> CreateRecipes(string mkDocsFilePath,
+            Dictionary<string, string> labels,
+            bool buildContainer)
         {
             var workingDirectory = Path.GetDirectoryName(mkDocsFilePath)!;
             var requirementsFilePath = Path.Combine(
@@ -75,7 +88,7 @@ namespace Bake.Cooking.Composers
                     workingDirectory);
             }
 
-            var outputDirectory = Path.Combine(workingDirectory, "site");
+            var outputDirectory = $"bake-site-{DateTimeOffset.Now:yyyy-MM-dd-HHmmss-fff}";
 
             yield return new MkDocsBuildRecipe(
                 workingDirectory,
@@ -84,6 +97,20 @@ namespace Bake.Cooking.Composers
                 outputDirectory,
                 new DocumentationSiteArtifact(
                     outputDirectory));
+
+            var dockerFilePath = Path.Combine(
+                workingDirectory,
+                "Dockerfile");
+
+            if (buildContainer)
+            {
+                yield return new MkDocsDockerFileRecipe(
+                    outputDirectory,
+                    labels,
+                    new DockerFileArtifact(
+                        "docs",
+                        dockerFilePath));
+            }
         }
     }
 }